[skip ci] Splitting problem module into three and uniforming the treatment of enums
