+++ PyGMO/CMakeLists.txt
@@ -15,6 +15,7 @@ ENDIF(ENABLE_GTOP_DATABASE)
 # Definitions needed when using the PaGMO DLL.
 SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${DYNAMIC_LIB_PAGMO_USE_FLAGS}")
 
+# This MACRO accepts a module name (i.e. _problem) and creates the target library and install target
 MACRO(PYTHON_CPPMODULE arg1)
 	SET_TARGET_PROPERTIES(${arg1} PROPERTIES PREFIX "")
 	IF(PYDEXTENSION)
@@ -29,6 +30,22 @@ MACRO(PYTHON_CPPMODULE arg1)
 		LIBRARY DESTINATION ${PYGMO_INSTALL_PATH}/${DIR_ARG}
 	)
 ENDMACRO(PYTHON_CPPMODULE)
+
+# This MACRO is the same as above, but allows to set the install directory manually (useful when splitting the boost_python modules in more files)
+MACRO(PYTHON_CPPMODULE_ADDITION arg1 arg2)
+	SET_TARGET_PROPERTIES(${arg1} PROPERTIES PREFIX "")
+	IF(PYDEXTENSION)
+		SET_TARGET_PROPERTIES(${arg1} PROPERTIES SUFFIX ".pyd")
+	ELSEIF(SOEXTENSION)
+		SET_TARGET_PROPERTIES(${arg1} PROPERTIES SUFFIX ".so")
+	ENDIF(PYDEXTENSION)
+	TARGET_LINK_LIBRARIES(${arg1} ${ARGN} ${MANDATORY_LIBRARIES} ${PYTHON_LIBRARY})
+	STRING(REGEX REPLACE "^_" "" DIR_ARG ${arg1})
+	INSTALL(TARGETS ${arg1}
+		RUNTIME DESTINATION ${PYGMO_INSTALL_PATH}/${arg2}
+		LIBRARY DESTINATION ${PYGMO_INSTALL_PATH}/${arg2}
+	)
+ENDMACRO(PYTHON_CPPMODULE_ADDITION)
 
 ADD_SUBDIRECTORY(algorithm)
 ADD_SUBDIRECTORY(core)
