+++ PyGMO/problem/problem.cpp
@@ -44,11 +44,6 @@
 #include "python_base.h"
 #include "python_base_stochastic.h"
 
-#ifdef PAGMO_ENABLE_KEP_TOOLBOX
-		#include <keplerian_toolbox/planets/base.h>
-		#include "../../src/problem/tsp_ds.h"
-#endif
-
 using namespace boost::python;
 using namespace pagmo;
 
@@ -60,40 +55,6 @@ static inline tuple find_subsequence_wrapper_cs(const problem::tsp_cs& p, const 
 	decision_vector::size_type retval_it_l, retval_it_r;
     p.find_subsequence(tour,retval_p,retval_l,retval_it_l,retval_it_r);
 	return boost::python::make_tuple(retval_p,retval_l,retval_it_l,retval_it_r);
-}
-
-#ifdef PAGMO_ENABLE_KEP_TOOLBOX
-	// wrapper for the find_subsequence method of tsp_ads
-	static inline tuple find_subsequence_wrapper_ds(const problem::tsp_ds& p, const decision_vector& tour, const bool static_computations)
-	{
-		double retval_p, retval_l;
-		decision_vector::size_type retval_it_l, retval_it_r;
-	    p.find_subsequence(tour,retval_p,retval_l,retval_it_l,retval_it_r, static_computations);
-		return boost::python::make_tuple(retval_p,retval_l,retval_it_l,retval_it_r);
-	}
-#endif
-
-// Transforms an Eigen Matrix into a std::vector<std::vector<double> >
-std::vector<std::vector<double> > get_rotation_matrix_from_eigen(const problem::rotated & p) {
-	Eigen::MatrixXd rot = p.get_rotation_matrix();
-	pagmo_assert(rot.cols()==rot.rows());
-	size_t dim = rot.cols();
-	std::vector<double> dummy(dim,0);
-	std::vector<std::vector<double> > retval(dim,dummy);
-
-	for (size_t i=0; i<dim;++i){
-		for (size_t j=0; j<dim;++j){
-			retval[i][j] =rot(i,j);
-		}
-	}
-	return retval;
-}
-
-// wrapper of a decompose method
-static inline fitness_vector compute_decomposed_fitness_wrapper(const problem::decompose& p, const fitness_vector &original_fit, const fitness_vector &weights) {
-	fitness_vector retval(1);
-	p.compute_decomposed_fitness(retval,original_fit,weights);
-	return retval;
 }
 
 // Wrapper to expose problems.
@@ -123,20 +84,6 @@ static inline class_<Problem,bases<problem::base>,bases<problem::base_stochastic
 	retval.def("cpp_dumps", &py_cpp_dumps<Problem>);
 	retval.add_property("seed",&problem::base_stochastic::get_seed,&problem::base_stochastic::set_seed,
 		"Random seed used in the objective function evaluation.");
-	return retval;
-}
-
-// Wrapper to expose meta problems.
-template <class Problem>
-static inline class_<Problem,bases<problem::base>,bases<problem::base_meta> > meta_problem_wrapper(const char *name, const char *descr)
-{
-	class_<Problem,bases<problem::base>,bases<problem::base_meta> > retval(name,descr,init<const Problem &>());
-	retval.def(init<>());
-	retval.def("__copy__", &Py_copy_from_ctor<Problem>);
-	retval.def("__deepcopy__", &Py_deepcopy_from_ctor<Problem>);
-	retval.def_pickle(python_class_pickle_suite<Problem>());
-	retval.def("cpp_loads", &py_cpp_loads<Problem>);
-	retval.def("cpp_dumps", &py_cpp_dumps<Problem>);
 	return retval;
 }
 
@@ -448,235 +395,6 @@ BOOST_PYTHON_MODULE(_problem) {
 	unc_mo_problem_wrapper<problem::zdt>("zdt","ZDT")
 		.def(init<optional<size_t, size_t> >());
 	
-	// Meta-problems
-
-	
-// Death penalty enums
-	enum_<problem::death_penalty::method_type>("_death_method_type")
-		.value("SIMPLE", problem::death_penalty::SIMPLE)
-		.value("KURI", problem::death_penalty::KURI)
-		.value("WEIGHTED", problem::death_penalty::WEIGHTED);
-
-	// Death penalty meta-problem
-	meta_problem_wrapper<problem::death_penalty>("death_penalty","Constrained death penalty problem")
-		.def(init<optional<const problem::base &, problem::death_penalty::method_type, const std::vector<double> &> >());
-
-	// con2mo penalty enums
-	enum_<problem::con2mo::method_type>("_con2mo_method_type")
-		.value("OBJ_CSTRS", problem::con2mo::OBJ_CSTRS)
-		.value("OBJ_CSTRSVIO", problem::con2mo::OBJ_CSTRSVIO)
-		.value("OBJ_EQVIO_INEQVIO", problem::con2mo::OBJ_EQVIO_INEQVIO);
-
-	// Constrained to multi-objective meta-problem.
-	meta_problem_wrapper<problem::con2mo>("con2mo","Constrained to multi-objective problem")
-		.def(init<optional<const problem::base &, problem::con2mo::method_type> >());
-
-	// con2uncon penalty enums
-	enum_<problem::con2uncon::method_type>("_con2uncon_method_type")
-		.value("OPTIMALITY", problem::con2uncon::OPTIMALITY)
-		.value("FEASIBILITY", problem::con2uncon::FEASIBILITY);
-
-	// Constrained to unconstrained meta-problem.
-	meta_problem_wrapper<problem::con2uncon>("con2uncon","Constrained to unconstrained problem")
-		.def(init<optional<const problem::base &, problem::con2uncon::method_type> >());
-
-	// Shifted meta-problem
-	meta_problem_wrapper<problem::shifted>("shifted","Shifted problem")
-		.def(init<const problem::base &>())
-		.def(init<const problem::base &, std::vector<double> >())
-		.def(init<const problem::base &, double>())
-		.add_property("shift_vector",make_function(&problem::shifted::get_shift_vector,return_value_policy<copy_const_reference>()))
-		.def("deshift",&problem::shifted::deshift);
-		
-	// Scaled meta-problem
-	meta_problem_wrapper<problem::scaled>("scaled","Scaled problem")
-		.def(init<const problem::base &, fitness_vector >())
-		.add_property("units",make_function(&problem::scaled::get_units,return_value_policy<copy_const_reference>()))
-		.def("descale",&problem::scaled::descale);
-		
-	// Rotated meta-problem
-	meta_problem_wrapper<problem::rotated>("rotated","Rotated problem")
-		.def(init<const problem::base &>())
-		.def(init<const problem::base &, Eigen::MatrixXd >())
-		.add_property("rotation_matrix",&get_rotation_matrix_from_eigen)
-		.def("derotate",&problem::rotated::derotate);
-		
-	// Normalized meta-problem
-	meta_problem_wrapper<problem::normalized>("normalized","Normalized problem")
-		.def(init<const problem::base &>())
-		.def("denormalize", &problem::normalized::denormalize);
-
-	// Decomposition meta-problem
-	// Exposing enums of problem::decompose
-	enum_<problem::decompose::method_type>("_decomposition_method")
-		.value("WEIGHTED", problem::decompose::WEIGHTED)
-		.value("TCHEBYCHEFF", problem::decompose::TCHEBYCHEFF)
-		.value("BI", problem::decompose::BI);
-	meta_problem_wrapper<problem::decompose>("decompose","Decomposed problem")
-		.def(init<const problem::base &, optional<problem::decompose::method_type, const std::vector<double> &, const std::vector<double> &, const bool> >())
-		.def("compute_decomposed_fitness", &compute_decomposed_fitness_wrapper, 
-		"Computes the fitness of the decomposed problem\n\n"
-		"  USAGE:: w = prob.compute_decomposed_fitness(fit,weight)\n"
-		"   - fit: multi-dimensional fitness\n"
-		"   - weight: decomposition weights")
-		.add_property("weights", make_function(&problem::decompose::get_weights, return_value_policy<copy_const_reference>()))
-		.add_property("ideal_point", &problem::decompose::get_ideal_point, &problem::decompose::set_ideal_point,"the (z) point used to compute tchebycheff and bi decompositions");
-
-
-	// Noisy meta-problem
-	// Exposing enums of problem::noisy
-	enum_<problem::noisy::noise_type>("_noise_distribution")
-		.value("NORMAL", problem::noisy::NORMAL)
-		.value("UNIFORM", problem::noisy::UNIFORM);
-
-	stochastic_problem_wrapper<problem::noisy>("noisy", "Noisy problem")
-		.def(init<const problem::base &, unsigned int, const double, const double, problem::noisy::noise_type, unsigned int>())
-		.add_property("noise_param_first", &problem::noisy::get_param_first)
-		.add_property("noise_param_second", &problem::noisy::get_param_second);
-
-	// Robust meta-problem
-	stochastic_problem_wrapper<problem::robust>("robust", "Robust problem")
-		.def(init<const problem::base &,unsigned int, const double, unsigned int>())
-		.add_property("rho", &problem::robust::get_rho);
-
-
-#ifdef PAGMO_ENABLE_KEP_TOOLBOX
-	// Asteroid Sample Return (also used fot human missions to asteroids)
-//	problem_wrapper<problem::sample_return>("sample_return","Asteroid sample return problem.")
-//		.def(init<const ::kep_toolbox::planet &, optional<const double &> >())
-//		.def("get_delta_v",&problem::sample_return::get_delta_v);
-
-	// Earth-planet problem.
-	//problem_wrapper<problem::earth_planet>("earth_planet","Earth-planet low-thrust problem.")
-		//.def(init< optional<int, std::string, const double &> >());
-
-	// GTOC5 launch.
-	//problem_wrapper<problem::gtoc5_launch>("gtoc5_launch","GTOC5 launch phase.")
-		//.def(init< optional<int, int, problem::gtoc5_launch::objective, const double &> >());
-
-	//enum_<problem::gtoc5_launch::objective>("gtoc5_launch_objective")
-		//.value("MASS",problem::gtoc5_launch::MASS)
-		//.value("TIME",problem::gtoc5_launch::TIME);
-
-	// GTOC5 randez-vouz.
-	//problem_wrapper<problem::gtoc5_rendezvous>("gtoc5_rendezvous","GTOC5 rendezvous phase.")
-		//.def(init< optional<int, int, int, const double &, const double &, const double &> >());
-
-	//enum_<problem::gtoc5_flyby::objective>("gtoc5_flyby_objective")
-		//.value("MASS",problem::gtoc5_flyby::MASS)
-		//.value("FINAL_EPOCH",problem::gtoc5_flyby::FINAL_EPOCH)
-		//.value("TIME",problem::gtoc5_flyby::TIME);
-
-	// GTOC5 flyby.
-	//problem_wrapper<problem::gtoc5_flyby>("gtoc5_flyby","GTOC5 flyby phase.")
-		//.def(init< optional<int, int, int, int, const double &, const double &, problem::gtoc5_flyby::objective, const double &, const double &> >());
-
-	// GTOC5 self flyby.
-	//problem_wrapper<problem::gtoc5_self_flyby>("gtoc5_self_flyby","GTOC5 self flyby phase.")
-		//.def(init< optional<int, int, const double &, const double &, const double &> >());
-
-	// GTOC1 problem.
-	problem_wrapper<problem::gtoc_1>("gtoc_1","GTOC 1 problem (chemical approximation).");
-
-	// GTOC2 problem.
-	problem_wrapper<problem::gtoc_2>("gtoc_2","GTOC 2 problem (LT model).")
-		.def(init<optional<int,int,int,int,int,problem::gtoc_2::objective> >());
-
-	// GTOC2's objectives enum.
-	enum_<problem::gtoc_2::objective>("_gtoc_2_objective")
-		.value("MASS",problem::gtoc_2::MASS)
-		.value("TIME",problem::gtoc_2::TIME)
-		.value("MASS_TIME",problem::gtoc_2::MASS_TIME);
-
-	// Laplace problem.
-	problem_wrapper<problem::laplace>("laplace","Laplace problem.")
-		.def(init< const std::vector<int> &>());
-
-	// Cassini 1.
-	problem_wrapper<problem::cassini_1>("cassini_1","Cassini 1 interplanetary trajectory problem.")
-		.def(init<optional<unsigned int> >());
-	// Messenger full.
-	problem_wrapper<problem::messenger_full>("messenger_full","Full Messenger problem.");
-
-	// Cassini 2.
-	problem_wrapper<problem::cassini_2>("cassini_2","Cassini 2 interplanetary trajectory problem.");
-
-	// Rosetta problem.
-	problem_wrapper<problem::rosetta>("rosetta","Rosetta problem.");
-
-	// Sagas problem.
-	problem_wrapper<problem::sagas>("sagas","Sagas problem.");
-
-	// Tandem.
-	problem_wrapper<problem::tandem>("tandem","Tandem problem.")
-		.def(init< optional<int, double> >())
-		.def("pretty", &problem::tandem::pretty);
-
-	problem_wrapper<problem::mga_1dsm_alpha>("mga_1dsm_alpha", "A Multiple Gravity Assist with 1 Deep Space Manouvre problem")
-		.def(init< optional<std::vector<kep_toolbox::planets::planet_ptr>, kep_toolbox::epoch, kep_toolbox::epoch, double, double, double, double, bool, bool, bool> >())
-		.def("pretty", &problem::mga_1dsm_alpha::pretty)
-		.def("set_tof", &problem::mga_1dsm_alpha::set_tof)
-		.def("get_tof", &problem::mga_1dsm_alpha::get_tof)
-		.def("set_launch_window", &problem::mga_1dsm_alpha::set_launch_window)
-		.def("set_vinf", &problem::mga_1dsm_alpha::set_vinf)
-		.def("get_sequence", &problem::mga_1dsm_alpha::get_sequence);
-
-	problem_wrapper<problem::mga_1dsm_tof>("mga_1dsm_tof", "A Multiple Gravity Assist with 1 Deep Space Manouvre problem")
-		.def(init< optional<std::vector<kep_toolbox::planets::planet_ptr>, kep_toolbox::epoch, kep_toolbox::epoch, std::vector<boost::array<double,2> >, double, double, bool, bool, bool> >())
-		.def("pretty", &problem::mga_1dsm_tof::pretty, (arg("x"),arg("extended_output") = false))
-		.def("set_tof", &problem::mga_1dsm_tof::set_tof)
-		.def("get_tof", &problem::mga_1dsm_tof::get_tof)
-		.def("set_launch_window", &problem::mga_1dsm_tof::set_launch_window)
-		.def("set_vinf", &problem::mga_1dsm_tof::set_vinf)
-		.def("get_sequence", &problem::mga_1dsm_tof::get_sequence);
-		
-	problem_wrapper<problem::mga_incipit>("mga_incipit", "Jupiter capture problem from the first part of gtoc6")
-		.def(init< optional< std::vector<kep_toolbox::planets::planet_ptr>, kep_toolbox::epoch, kep_toolbox::epoch, std::vector<std::vector<double> > > >())
-		.def("pretty", &problem::mga_incipit::pretty)
-		.def("get_sequence", &problem::mga_incipit::get_sequence)
-		.add_property("tof",make_function(&problem::mga_incipit::get_tof, return_value_policy<copy_const_reference>()), &problem::mga_incipit::set_tof,"bound on the times of flight for the different legs");
-
-	problem_wrapper<problem::mga_incipit_cstrs>("mga_incipit_cstrs", "Jupiter capture problem from the first part of gtoc6 (constrained version)")
-		.def(init< optional< std::vector<kep_toolbox::planets::planet_ptr>, kep_toolbox::epoch, kep_toolbox::epoch, std::vector<std::vector<double> >, const double, const double > >())
-		.def("pretty", &problem::mga_incipit_cstrs::pretty)
-		.def("get_sequence", &problem::mga_incipit_cstrs::get_sequence)
-		.add_property("tof",make_function(&problem::mga_incipit_cstrs::get_tof, return_value_policy<copy_const_reference>()), &problem::mga_incipit_cstrs::set_tof,"bound on the times of flight for the different legs");
-
-	problem_wrapper<problem::mga_part>("mga_part", "A part of the Jupiter moon tour from gtoc6")
-		.def(init< optional <std::vector<kep_toolbox::planets::planet_ptr>, std::vector<std::vector<double> >, kep_toolbox::epoch, kep_toolbox::array3D > >())
-		.def("pretty", &problem::mga_part::pretty)
-		.def("get_sequence", &problem::mga_part::get_sequence)
-		.add_property("vinf_in",make_function(&problem::mga_part::get_vinf_in, return_value_policy<copy_const_reference>()), &problem::mga_part::set_vinf_in,"initial incoming relative spacecraft velocity")
-		.add_property("t0",make_function(&problem::mga_part::get_t0, return_value_policy<copy_const_reference>()), &problem::mga_part::set_t0, "start epoch")
-		.add_property("tof",make_function(&problem::mga_part::get_tof, return_value_policy<copy_const_reference>()), &problem::mga_part::set_tof,"bounds on the times of flight for the different legs")
-		.add_property("betas",&problem::mga_part::get_betas, &problem::mga_part::set_betas,"bounds on the beta angles for the different legs")
-		.add_property("rps",&problem::mga_part::get_rps, &problem::mga_part::set_rps,"bounds on the periplanet heights for the different legs");
-
-	// Travelling salesman problem, Asteroids / Debris Selection TSP (TSP-ADS)
-	tsp_problem_wrapper<problem::tsp_ds>("tsp_ds","Debris Selection TSP (TSP-DS)")
-		.def(init<optional<const std::vector<kep_toolbox::planets::planet_ptr>&, const std::vector<double>&, const double, const std::vector<double>&, const problem::base_tsp::encoding_type & > >())
-		.def("find_subsequence", &find_subsequence_wrapper_ds)
-        .add_property("planets", make_function(&problem::tsp_ds::get_planets, return_value_policy<copy_const_reference>()) )
-        .add_property("values", make_function(&problem::tsp_ds::get_values, return_value_policy<copy_const_reference>()) )
-        .add_property("epochs", make_function(&problem::tsp_ds::get_epochs, return_value_policy<copy_const_reference>()), "epoch schedule")
-        .add_property("max_DV", &problem::tsp_ds::get_max_DV );
-#endif
-
-#ifdef PAGMO_ENABLE_GSL
-	// Spheres Problems
-	stochastic_problem_wrapper<problem::spheres>("mit_spheres", "Spheres problem, a neurocontroller for the MIT test-bed (absolute perception-action)")
-		.def(init< optional<int,int,double,unsigned int, bool, double, std::vector<double> > >())
-		.def("post_evaluate", &problem::spheres::post_evaluate)
-		.def("simulate", &problem::spheres::simulate)
-		.def("get_nn_weights", &problem::spheres::get_nn_weights)
-		.add_property("seed",&problem::spheres::get_seed,&problem::spheres::set_seed,"Random seed used in the objective function evaluation.");
-
-	//problem_wrapper<problem::spheres_q>("spheres_q", "Spheres problem, a neurocontroller for the MIT test-bed (body-axis perception-action)")
-	//	.def(init< optional<int,int,double,unsigned int> >())
-	//	.def("post_evaluate", &problem::spheres_q::post_evaluate)
-	//	.def("simulate", &problem::spheres_q::simulate);
-#endif
-
 	// Register to_python conversion from smart pointer.
 	register_ptr_to_python<problem::base_ptr>();
 }
